# doc/client_gateway_service/v1/swagger.yaml
basePath: /client-gateway/v1
definitions:
...
  models.User:
    properties:
      email:
        type: string
      email_verified:
        type: boolean
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      phone_verified:
        type: boolean
      photo_path:
        type: string
      reg_datetime:
        type: string
      user_id:
        type: integer
    type: object
...
  schema.ApprovePhone:
    properties:
      code:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    type: object
...
  schema.ReqLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  schema.ReqRefreshToken:
    properties:
      refresh_token:
        type: string
    type: object
  schema.ReqSignUp:
    properties:
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  schema.RespLogin:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  schema.RespRefreshToken:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact: {}
  description: Gateway service for simple users
  title: Client gateway service
  version: "1.0"
paths:
  /auth/approve-phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: ApprovePhone data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/schema.ApprovePhone'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Подтверждение номера телефона
      tags:
      - Авторизация
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/schema.ReqLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.RespLogin'
      summary: Логин
      tags:
      - Авторизация
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Деавторизация
      tags:
      - Авторизация
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/schema.ReqRefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.RespRefreshToken'
      summary: Обновление токена
      tags:
      - Авторизация
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Signup data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/schema.ReqSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Регистрация пользователя
      tags:
      - Авторизация
...
swagger: "2.0"